Add TableCollectors to AndroidAccessToCollectors
RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=342328137
