+++ guava/src/com/google/common/collect/Tables.java
@@ -68,14 +68,7 @@ public final class Tables {
       java.util.function.Function<? super T, ? extends C> columnFunction,
       java.util.function.Function<? super T, ? extends V> valueFunction,
       java.util.function.Supplier<I> tableSupplier) {
-    return toTable(
-        rowFunction,
-        columnFunction,
-        valueFunction,
-        (v1, v2) -> {
-          throw new IllegalStateException("Conflicting values " + v1 + " and " + v2);
-        },
-        tableSupplier);
+    return TableCollectors.toTable(rowFunction, columnFunction, valueFunction, tableSupplier);
   }
 
   /**
@@ -98,42 +91,8 @@ public final class Tables {
       java.util.function.Function<? super T, ? extends V> valueFunction,
       BinaryOperator<V> mergeFunction,
       java.util.function.Supplier<I> tableSupplier) {
-    checkNotNull(rowFunction);
-    checkNotNull(columnFunction);
-    checkNotNull(valueFunction);
-    checkNotNull(mergeFunction);
-    checkNotNull(tableSupplier);
-    return Collector.of(
-        tableSupplier,
-        (table, input) ->
-            merge(
-                table,
-                rowFunction.apply(input),
-                columnFunction.apply(input),
-                valueFunction.apply(input),
-                mergeFunction),
-        (table1, table2) -> {
-          for (Table.Cell<R, C, V> cell2 : table2.cellSet()) {
-            merge(table1, cell2.getRowKey(), cell2.getColumnKey(), cell2.getValue(), mergeFunction);
-          }
-          return table1;
-        });
-  }
-
-  private static <R, C, V> void merge(
-      Table<R, C, V> table, R row, C column, V value, BinaryOperator<V> mergeFunction) {
-    checkNotNull(value);
-    V oldValue = table.get(row, column);
-    if (oldValue == null) {
-      table.put(row, column, value);
-    } else {
-      V newValue = mergeFunction.apply(oldValue, value);
-      if (newValue == null) {
-        table.remove(row, column);
-      } else {
-        table.put(row, column, newValue);
-      }
-    }
+    return TableCollectors.toTable(
+        rowFunction, columnFunction, valueFunction, mergeFunction, tableSupplier);
   }
 
   /**
